🌱 Split server_operations.go into domain-specific files
Delete the 2158-line pkg/agent/server_operations.go monolith and
replace it with 4 focused domain files:

- server_ops_settings.go: Settings CRUD, API key validation per
  provider (Claude, OpenAI, OpenRouter, Groq, Gemini), provider
  health checks (956 lines)
- server_ops_clusters.go: Cloud CLI detection, local cluster
  tools/lifecycle, vCluster CRUD (642 lines)
- server_ops_predictions.go: AI predictions, analyze, feedback,
  stats, metrics history, device alerts/inventory (454 lines)
- server_ops_insights.go: Insights enrichment, insights AI,
  vCluster availability check (140 lines)

All methods remain on *Server in the same package — zero interface
or behavior changes.

Fixes #10411
